MULE-8080: Oracle Configuration still requires user and password attributes when it is configured via URL
